use torch.float32 to express angle

+++ mmseg/datasets/transforms/formatting.py
@@ -219,7 +219,7 @@ class PackInstanceSegInputs(BaseTransform):
             direction_angle = np.arctan2(direction_y, direction_x)
             
             direction_data = dict(
-                data=to_tensor(direction_angle[None, ...])) # adding None to add a new axis for batch size
+                data=to_tensor(direction_angle[None, ...].astype(np.float32))) # adding None to add a new axis for batch size
 
             data_sample.set_data(
                 dict(direction_map=PixelData(**direction_data)))
